refactor: convert frontend components from TSX to JSX
- Refactor: Stripped all TypeScript interfaces, types, and inline annotations from Home screen and related components.
- Chore: Renamed files from `.tsx` to `.jsx` to unify the frontend codebase language.
- Affected: Feed, Reels, Stories, Create Post, Comments, and Navigation layouts.
