+++ mobile-app/app/create-post.jsx
@@ -1,16 +1,13 @@
 /**
  * ====================================
- * CREATE REEL SCREEN
+ * CREATE POST SCREEN
  * ====================================
- * Modal-card style reel creation with:
- * - Video upload (required) via expo-image-picker
- * - Caption input
+ * Modal-card style post creation with:
+ * - Caption text input
+ * - Media picker (images only) via expo-image-picker
  * - Category selector with brand-red highlight
- * - Tags input
- * - Direct "Share Reel" upload button
- * - Preview navigation button
- *
- * Matches the Create Post screen design.
+ * - Image preview with clear button
+ * - Navigation to preview-post screen
  */
 
 import React, { useState } from "react";
@@ -19,6 +16,7 @@ import {
   Text,
   TextInput,
   TouchableOpacity,
+  Image,
   Alert,
   StyleSheet,
   ScrollView,
@@ -30,7 +28,6 @@ import {
 import { useRouter } from "expo-router";
 import { Ionicons } from "@expo/vector-icons";
 import * as ImagePicker from "expo-image-picker";
-import { Video, ResizeMode } from "expo-av";
 import { useAuth } from "../src/context/AuthContext";
 import { POST_ENDPOINTS } from "../src/config/api";
 
@@ -46,15 +43,16 @@ const CATEGORIES = [
   "Memes",
 ];
 
-export default function CreateReelScreen() {
+export default function CreatePostNewScreen() {
   const router = useRouter();
   const { user } = useAuth();
 
   // ── Form state ──
-  const [selectedVideo, setSelectedVideo] = useState<string | null>(null);
   const [caption, setCaption] = useState("");
+  const [media, setMedia] = useState(null);
   const [category, setCategory] = useState("General");
   const [tags, setTags] = useState("");
+  const [isUploading, setIsUploading] = useState(false);
   const [isSubmitting, setIsSubmitting] = useState(false);
 
   // ====================================
@@ -67,13 +65,13 @@ export default function CreateReelScreen() {
           <Text style={styles.guardIcon}>🔒</Text>
           <Text style={styles.guardTitle}>Not Logged In</Text>
           <Text style={styles.guardSubtitle}>
-            Please log in to create a reel
+            Please log in to create a post
           </Text>
           <TouchableOpacity
-            style={styles.shareButton}
+            style={styles.previewButton}
             onPress={() => router.replace("/")}
           >
-            <Text style={styles.shareButtonText}>Go to Login</Text>
+            <Text style={styles.previewButtonText}>Go to Login</Text>
           </TouchableOpacity>
         </View>
       </SafeAreaView>
@@ -81,9 +79,9 @@ export default function CreateReelScreen() {
   }
 
   // ====================================
-  // PICK VIDEO FROM GALLERY
+  // PICK MEDIA FROM GALLERY
   // ====================================
-  const pickVideo = async () => {
+  const pickMedia = async () => {
     try {
       const permissionResult =
         await ImagePicker.requestMediaLibraryPermissionsAsync();
@@ -97,61 +95,67 @@ export default function CreateReelScreen() {
       }
 
       const result = await ImagePicker.launchImageLibraryAsync({
-        mediaTypes: ImagePicker.MediaTypeOptions.Videos,
+        mediaTypes: ImagePicker.MediaTypeOptions.Images,
         allowsEditing: true,
         quality: 0.8,
-        videoMaxDuration: 60, // 60 seconds max
       });
 
       if (!result.canceled && result.assets[0]) {
-        setSelectedVideo(result.assets[0].uri);
-        console.log(`📎 Video selected: ${result.assets[0].uri.slice(-30)}`);
+        const asset = result.assets[0];
+        setMedia(asset.uri);
+
+        console.log(`📎 Image selected: ${asset.uri.slice(-30)}`);
       }
     } catch (error) {
-      console.error("Video pick error:", error);
-      Alert.alert("Error", "Failed to pick video. Please try again.");
+      console.error("Media pick error:", error);
+      Alert.alert("Error", "Failed to pick media. Please try again.");
     }
   };
 
   // ====================================
-  // CLEAR SELECTED VIDEO
+  // CLEAR SELECTED MEDIA
   // ====================================
-  const clearVideo = () => {
-    setSelectedVideo(null);
+  const clearMedia = () => {
+    setMedia(null);
   };
 
   // ====================================
   // HANDLE PREVIEW NAVIGATION
   // ====================================
   const handlePreview = () => {
-    if (!selectedVideo) {
-      Alert.alert("Required Field", "Please select a video before continuing.");
+    if (!caption.trim() && !media) {
+      Alert.alert(
+        "Required",
+        "Please add a caption or select media before continuing.",
+      );
       return;
     }
 
-    // Navigate to preview screen with reel data
+    // Navigate to preview screen with post data
     router.push({
-      pathname: "/preview-reel",
+      pathname: "/preview-post",
       params: {
-        video: selectedVideo,
-        caption: caption,
+        content: caption,
+        media: media || "",
+        mediaType: "image",
+        category: category,
         tags: tags,
       },
     });
   };
 
   // ====================================
-  // HANDLE SHARE REEL (UPLOAD)
+  // HANDLE SHARE POST (UPLOAD)
   // ====================================
-  const handleShareReel = async () => {
+  const handleSharePost = async () => {
     // Validate
     if (!user) {
       Alert.alert("Error", "You must be logged in to post.");
       return;
     }
 
-    if (!selectedVideo) {
-      Alert.alert("Video Required", "Please select a video to share.");
+    if (!media) {
+      Alert.alert("Media Required", "Please select a photo to share.");
       return;
     }
 
@@ -167,17 +171,17 @@ export default function CreateReelScreen() {
       // CRITICAL ANDROID FIX:
       // Android requires the file to be appended as an object with
       // uri, name, and type — not as a Blob.
-      const filename = selectedVideo.split("/").pop() || "upload.mp4";
+      const filename = media.split("/").pop() || "upload.jpg";
       const match = /\.(\w+)$/.exec(filename);
-      const type = match ? `video/${match[1]}` : "video/mp4";
+      const type = match ? `image/${match[1]}` : "image/jpeg";
 
       formData.append("media", {
-        uri: selectedVideo,
+        uri: media,
         name: filename,
         type: type,
-      } as any);
+      });
 
-      console.log(`📤 Uploading video: ${filename} (${type})`);
+      console.log(`📤 Uploading image: ${filename} (${type})`);
 
       // POST to backend
       const response = await fetch(POST_ENDPOINTS.CREATE, {
@@ -191,13 +195,13 @@ export default function CreateReelScreen() {
       const data = await response.json();
 
       if (data.success) {
-        Alert.alert("🎉 Success!", "Your reel has been shared!", [
+        Alert.alert("🎉 Success!", "Your post has been shared!", [
           {
             text: "OK",
             onPress: () => {
               // Clear all states
-              setSelectedVideo(null);
               setCaption("");
+              setMedia(null);
               setCategory("General");
               setTags("");
 
@@ -239,7 +243,7 @@ export default function CreateReelScreen() {
         <View style={styles.card}>
           {/* Header */}
           <View style={styles.header}>
-            <Text style={styles.headerTitle}>Create Reel</Text>
+            <Text style={styles.headerTitle}>Create Post</Text>
             <TouchableOpacity
               style={styles.closeButton}
               onPress={() => router.back()}
@@ -250,10 +254,12 @@ export default function CreateReelScreen() {
 
           {/* ========== CAPTION INPUT ========== */}
           <View style={styles.inputSection}>
-            <Text style={styles.label}>Caption</Text>
+            <Text style={styles.label}>
+              What's on your mind? <Text style={styles.required}>*</Text>
+            </Text>
             <TextInput
               style={styles.textArea}
-              placeholder="Write a caption..."
+              placeholder="Share your thoughts..."
               placeholderTextColor="#999"
               value={caption}
               onChangeText={setCaption}
@@ -263,41 +269,28 @@ export default function CreateReelScreen() {
             />
           </View>
 
-          {/* ========== VIDEO PICKER ========== */}
+          {/* ========== MEDIA PICKER (Images Only) ========== */}
           <View style={styles.inputSection}>
-            <Text style={styles.label}>
-              Video <Text style={styles.required}>*</Text>
-            </Text>
-            <TouchableOpacity style={styles.mediaPicker} onPress={pickVideo}>
-              {selectedVideo ? (
+            <Text style={styles.label}>Photo (Optional)</Text>
+            <TouchableOpacity style={styles.mediaPicker} onPress={pickMedia}>
+              {media ? (
                 <View style={styles.mediaPreviewContainer}>
-                  <Video
-                    source={{ uri: selectedVideo }}
-                    style={styles.mediaPreview}
-                    resizeMode={ResizeMode.COVER}
-                    shouldPlay={false}
-                    isLooping={false}
-                    isMuted
-                  />
-                  {/* Play icon overlay */}
-                  <View style={styles.videoOverlay}>
-                    <Ionicons name="play-circle" size={48} color="#fff" />
-                  </View>
+                  <Image source={{ uri: media }} style={styles.mediaPreview} />
 
                   {/* X button to clear selection */}
                   <TouchableOpacity
                     style={styles.removeMedia}
-                    onPress={clearVideo}
+                    onPress={clearMedia}
                   >
                     <Ionicons name="close-circle" size={26} color="#f9252b" />
                   </TouchableOpacity>
                 </View>
               ) : (
                 <View style={styles.mediaPlaceholder}>
-                  <Ionicons name="videocam-outline" size={32} color="#999" />
-                  <Text style={styles.mediaText}>Add a video</Text>
+                  <Ionicons name="images-outline" size={32} color="#999" />
+                  <Text style={styles.mediaText}>Add a photo</Text>
                   <Text style={styles.mediaSubtext}>
-                    Tap to select from gallery (max 60s)
+                    Tap to select from gallery
                   </Text>
                 </View>
               )}
@@ -339,7 +332,7 @@ export default function CreateReelScreen() {
             <Text style={styles.label}>Tags</Text>
             <TextInput
               style={styles.input}
-              placeholder="music, dance, fun"
+              placeholder="campus, events, community"
               placeholderTextColor="#999"
               value={tags}
               onChangeText={setTags}
@@ -369,14 +362,14 @@ export default function CreateReelScreen() {
             </TouchableOpacity>
           </View>
 
-          {/* ========== SHARE REEL BUTTON ========== */}
+          {/* ========== SHARE POST BUTTON ========== */}
           <TouchableOpacity
             style={[
               styles.shareButton,
-              (!selectedVideo || isSubmitting) && styles.buttonDisabled,
+              (!media || isSubmitting) && styles.buttonDisabled,
             ]}
-            onPress={handleShareReel}
-            disabled={!selectedVideo || isSubmitting}
+            onPress={handleSharePost}
+            disabled={!media || isSubmitting}
             activeOpacity={0.8}
           >
             {isSubmitting ? (
@@ -387,7 +380,7 @@ export default function CreateReelScreen() {
             ) : (
               <View style={styles.shareButtonContent}>
                 <Ionicons name="paper-plane" size={18} color="#fff" />
-                <Text style={styles.shareButtonText}>Share Reel</Text>
+                <Text style={styles.shareButtonText}>Share Post</Text>
               </View>
             )}
           </TouchableOpacity>
@@ -477,7 +470,7 @@ const styles = StyleSheet.create({
     color: "#999",
     marginTop: 6,
   },
-  // Media Picker (matches Create Post)
+  // Media Picker
   mediaPicker: {
     borderWidth: 1,
     borderColor: "#e0e0e0",
@@ -506,16 +499,7 @@ const styles = StyleSheet.create({
   mediaPreview: {
     width: "100%",
     height: 200,
-  },
-  videoOverlay: {
-    position: "absolute",
-    top: 0,
-    left: 0,
-    right: 0,
-    bottom: 0,
-    backgroundColor: "rgba(0,0,0,0.3)",
-    alignItems: "center",
-    justifyContent: "center",
+    resizeMode: "cover",
   },
   removeMedia: {
     position: "absolute",
@@ -524,7 +508,7 @@ const styles = StyleSheet.create({
     backgroundColor: "#fff",
     borderRadius: 13,
   },
-  // Category Selector (matches Create Post)
+  // Category Selector
   categoryRow: {
     flexDirection: "row",
     flexWrap: "wrap",
@@ -551,7 +535,7 @@ const styles = StyleSheet.create({
     color: "#fff",
     fontWeight: "600",
   },
-  // Buttons (matches Create Post)
+  // Buttons
   buttonRow: {
     flexDirection: "row",
     justifyContent: "space-between",
@@ -583,12 +567,17 @@ const styles = StyleSheet.create({
     borderWidth: 1,
     borderColor: "#f9252b",
   },
+  previewButtonText: {
+    fontSize: 15,
+    fontWeight: "600",
+    color: "#fff",
+  },
   previewButtonTextOutline: {
     fontSize: 15,
     fontWeight: "600",
     color: "#f9252b",
   },
-  // Share Reel Button (matches Create Post)
+  // Share Post Button
   shareButton: {
     marginTop: 12,
     paddingVertical: 16,
