+++ mobile-app/app/preview-reel.jsx
@@ -1,9 +1,9 @@
 /**
  * ====================================
- * POST PREVIEW SCREEN
+ * REEL PREVIEW SCREEN
  * ====================================
- * Shows a preview of the post before publishing.
- * Displays user info, content, media, and tags.
+ * Shows a preview of the reel before publishing.
+ * Displays user info, video preview, caption, and tags.
  */
 
 import React, { useState } from "react";
@@ -22,17 +22,14 @@ import {
 } from "react-native";
 import { useRouter, useLocalSearchParams } from "expo-router";
 import { Ionicons } from "@expo/vector-icons";
+import { Video, ResizeMode } from "expo-av";
 import { useAuth } from "../src/context/AuthContext";
 import { POST_ENDPOINTS } from "../src/config/api";
 
-export default function PreviewPostScreen() {
+export default function PreviewReelScreen() {
   const router = useRouter();
   const { user } = useAuth();
-  const params = useLocalSearchParams<{
-    content: string;
-    media?: string;
-    tags?: string;
-  }>();
+  const params = useLocalSearchParams();
 
   const [isUploading, setIsUploading] = useState(false);
 
@@ -51,30 +48,32 @@ export default function PreviewPostScreen() {
       return;
     }
 
+    if (!params.video) {
+      Alert.alert("Error", "No video to upload.");
+      return;
+    }
+
     setIsUploading(true);
 
     try {
       const formData = new FormData();
-      formData.append("caption", params.content || "");
+      formData.append("caption", params.caption || "");
       formData.append("userId", user.id);
 
       if (params.tags) {
         formData.append("tags", params.tags);
       }
 
-      // Add media if present
-      if (params.media) {
-        const filename = params.media.split("/").pop() || "photo.jpg";
-        const match = /\.(\w+)$/.exec(filename);
-        const type = match ? `image/${match[1]}` : "image/jpeg";
+      // Add video media
+      const filename = params.video.split("/").pop() || "reel.mp4";
+      const match = /\.(\w+)$/.exec(filename);
+      const type = match ? `video/${match[1]}` : "video/mp4";
 
-        // @ts-ignore
-        formData.append("media", {
-          uri: params.media,
-          name: filename,
-          type: type,
-        });
-      }
+      formData.append("media", {
+        uri: params.video,
+        name: filename,
+        type: type,
+      });
 
       const response = await fetch(POST_ENDPOINTS.CREATE, {
         method: "POST",
@@ -84,11 +83,11 @@ export default function PreviewPostScreen() {
       const data = await response.json();
 
       if (data.success) {
-        Alert.alert("🎉 Success!", "Your post has been shared!", [
+        Alert.alert("🎉 Success!", "Your reel has been shared!", [
           {
             text: "OK",
             onPress: () => {
-              // Navigate back to home (go back multiple screens)
+              // Navigate back to home
               router.dismissAll();
               router.replace("/(tabs)");
             },
@@ -122,7 +121,7 @@ export default function PreviewPostScreen() {
         <View style={styles.card}>
           {/* Header */}
           <View style={styles.header}>
-            <Text style={styles.headerTitle}>Preview Your Post</Text>
+            <Text style={styles.headerTitle}>Preview Your Reel</Text>
             <TouchableOpacity
               style={styles.closeButton}
               onPress={() => router.back()}
@@ -131,8 +130,8 @@ export default function PreviewPostScreen() {
             </TouchableOpacity>
           </View>
 
-          {/* Post Preview Card */}
-          <View style={styles.postPreview}>
+          {/* Reel Preview Card */}
+          <View style={styles.reelPreview}>
             {/* User Info Row */}
             <View style={styles.userRow}>
               {user?.profilePicture ? (
@@ -151,16 +150,26 @@ export default function PreviewPostScreen() {
               </View>
             </View>
 
-            {/* Post Content */}
-            <Text style={styles.postContent}>{params.content}</Text>
+            {/* Video Preview */}
+            {params.video && (
+              <View style={styles.videoContainer}>
+                <Video
+                  source={{ uri: params.video }}
+                  style={styles.videoPreview}
+                  resizeMode={ResizeMode.COVER}
+                  shouldPlay={false}
+                  isLooping={false}
+                  isMuted
+                />
+                <View style={styles.videoOverlay}>
+                  <Ionicons name="play-circle" size={48} color="#fff" />
+                </View>
+              </View>
+            )}
 
-            {/* Media Preview (if any) */}
-            {params.media && (
-              <Image
-                source={{ uri: params.media }}
-                style={styles.mediaImage}
-                resizeMode="cover"
-              />
+            {/* Caption */}
+            {params.caption && (
+              <Text style={styles.caption}>{params.caption}</Text>
             )}
 
             {/* Tags */}
@@ -177,7 +186,7 @@ export default function PreviewPostScreen() {
 
           {/* Footer Text */}
           <Text style={styles.footerText}>
-            This is how your post will appear in the feed
+            This is how your reel will appear in the feed
           </Text>
 
           {/* Action Buttons */}
@@ -248,8 +257,8 @@ const styles = StyleSheet.create({
   closeButton: {
     padding: 4,
   },
-  // Post Preview
-  postPreview: {
+  // Reel Preview
+  reelPreview: {
     backgroundColor: "#fff",
     borderRadius: 12,
     borderWidth: 1,
@@ -293,18 +302,33 @@ const styles = StyleSheet.create({
     fontSize: 12,
     color: "#999",
   },
-  // Post Content
-  postContent: {
+  // Video
+  videoContainer: {
+    position: "relative",
+    borderRadius: 8,
+    overflow: "hidden",
+    marginBottom: 12,
+  },
+  videoPreview: {
+    width: "100%",
+    height: 200,
+    backgroundColor: "#000",
+  },
+  videoOverlay: {
+    position: "absolute",
+    top: 0,
+    left: 0,
+    right: 0,
+    bottom: 0,
+    backgroundColor: "rgba(0,0,0,0.3)",
+    alignItems: "center",
+    justifyContent: "center",
+  },
+  // Caption
+  caption: {
     fontSize: 15,
     color: "#262626",
     lineHeight: 22,
-    marginBottom: 12,
-  },
-  // Media
-  mediaImage: {
-    width: "100%",
-    height: 200,
-    borderRadius: 8,
     marginBottom: 12,
   },
   // Tags
