+++ mobile-app/package.json
@@ -42,10 +42,8 @@
     "react-native-worklets": "0.5.1"
   },
   "devDependencies": {
-    "@types/react": "~19.1.0",
     "eslint": "^9.25.0",
-    "eslint-config-expo": "~10.0.0",
-    "typescript": "~5.9.2"
+    "eslint-config-expo": "~10.0.0"
   },
   "private": true
 }
