--- mobile-app/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "compilerOptions": {},
-  "extends": "expo/tsconfig.base",
-  "include": [
-    "**/*.ts",
-    "**/*.tsx",
-    ".expo/types/**/*.ts",
-    "expo-env.d.ts"
-  ]
-}
